+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- **Breaking:** Changed `activity`, `artists` and `disclaimers` fields to be optional in TopArtistsResponse to prevent crashes
+
 ## [0.1.0] - 2025-01-15
 
 ### Added